made working HOT and SEARCH tabs - add specific request and add functions for them on Fragment and View Model
