+++ app/src/main/java/art/manguste/android/gamesearch/gamesList/GamesViewModel.kt
@@ -1,7 +1,6 @@
 package art.manguste.android.gamesearch.gamesList
 
 import android.app.Application
-import android.util.Log
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -10,7 +9,7 @@ import art.manguste.android.gamesearch.core.GameBriefly
 import art.manguste.android.gamesearch.network.GamesApi
 import kotlinx.coroutines.launch
 
-enum class ApiStatus { LOADING, ERROR, DONE }
+enum class LoadStatus { LOADING, ERROR, DONE }
 
 class GamesViewModel(application: Application) : AndroidViewModel(application) {
 
@@ -20,31 +19,70 @@ class GamesViewModel(application: Application) : AndroidViewModel(application) {
     private var _gamesList = MutableLiveData<List<GameBriefly>>()
     val gamesList: LiveData<List<GameBriefly>> get() = _gamesList
 
-    private val _status = MutableLiveData<ApiStatus>()
-    val status: LiveData<ApiStatus>
+    private val _status = MutableLiveData<LoadStatus>()
+    val status: LiveData<LoadStatus>
         get() = _status
 
     // displays data on init
     init {
-        getGamesHotListSearch()
+        //getGamesHotList()
     }
 
-    private fun getGamesHotListSearch() {
+    private fun getGamesHotList() {
         viewModelScope.launch {
             try {
-                _status.value = ApiStatus.LOADING
+                _status.value = LoadStatus.LOADING
                 val resultRequest = GamesApi.retrofitService.getGamesList()
                 _gamesList.value = resultRequest.results
                 //Log.d("GamesList fragment", "games = ${gamesList.value?.size}")
                 _response.value = "Got : ${resultRequest.next}  "//response.body()
-                _status.value = ApiStatus.DONE
+                _status.value = LoadStatus.DONE
             } catch (e: Exception) {
-                _status.value = ApiStatus.ERROR
+                _status.value = LoadStatus.ERROR
                 _response.value = "Not reachable : ${e.message}"
                 _gamesList.value = ArrayList()
             }
         }
+    }
 
+    fun getHotGamesList(){
+        viewModelScope.launch {
+            try {
+                _status.value = LoadStatus.LOADING
+
+                val resultRequest = GamesApi.retrofitService.getGamesHotListSorts(
+                        datesRange = "2020-06-01,2020-09-15",
+                        ordering = "-added")
+                // example: https://api.rawg.io/api/games?dates=2020-06-01,2020-09-15&ordering=-added
+
+                _gamesList.value = resultRequest.results
+                //Log.d("GamesList fragment", "games = ${gamesList.value?.size}")
+
+                _status.value = LoadStatus.DONE
+            } catch (e: Exception) {
+                _status.value = LoadStatus.ERROR
+                _gamesList.value = ArrayList()
+            }
+        }
+    }
+
+    fun getSearchGameList(search: String) {
+        viewModelScope.launch {
+            try {
+                _status.value = LoadStatus.LOADING
+
+                val resultRequest = GamesApi.retrofitService.getGamesListSearch(pageSize = 10,
+                        gameName = search)
+
+                _gamesList.value = resultRequest.results
+                //Log.d("GamesList fragment", "games = ${gamesList.value?.size}")
+
+                _status.value = LoadStatus.DONE
+            } catch (e: Exception) {
+                _status.value = LoadStatus.ERROR
+                _gamesList.value = ArrayList()
+            }
+        }
     }
 
 }
