+++ app/src/main/res/values/ids.xml
@@ -3,7 +3,7 @@
 
     <item name="ll_search" type="id"/>
     <item name="btnStartSearch" type="id"/>
-    <item name="et_search_by_name" type="id"/>
+    <item name="searchByTitle" type="id"/>
     <item name="progressBar" type="id"/>
     <item name="recyclerGames" type="id"/>
 </resources>
